refactor(web): new UI: 编辑器代码抽象
